Fix #212: Cleanup and modernize Browser sample
